Squashed commit of the following:
commit 08f2c40032efe2d5ba57cfb5c41612ae4ba90bca
Author: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Date:   Wed Oct 14 00:11:48 2020 +0000

    Bump junit from 4.12 to 4.13.1

    Bumps [junit](https://github.com/junit-team/junit4) from 4.12 to 4.13.1.
    - [Release notes](https://github.com/junit-team/junit4/releases)
    - [Changelog](https://github.com/junit-team/junit4/blob/main/doc/ReleaseNotes4.12.md)
    - [Commits](https://github.com/junit-team/junit4/compare/r4.12...r4.13.1)


commit 9d0e305151f29eaa72864f5bb49462f96bb49842
Author: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Date:   Wed Sep 16 02:50:49 2020 +0000

    Bump commons-fileupload from 1.3.1 to 1.3.3 in /blog2

    Bumps commons-fileupload from 1.3.1 to 1.3.3.

+++ blog2/src/main/java/com/blog/service/PictureService.java
@@ -18,13 +18,13 @@ public interface PictureService<T> {
     Picture query(Integer id);
 
     @ApiModelProperty("根据id修改")
-    long updatePictureById(T t);
+    int updatePictureById(T t);
 
     @ApiModelProperty("根据id删除")
-    long detelePictureById(Integer id);
+    int detelePictureById(Integer id);
 
     @ApiModelProperty("根据id添加")
-    long insert(T  t);
+    int insert(T  t);
 
     @ApiModelProperty("查询全部")
     List<T> BlogPicture();
